==> analysis/4.0_pcaExplore_GO.rmd <==
---
title: "PCAExpl.rmd"
output: html_document
author: Tina Gorsek Sparovec
date: 10.2.21
params:
 data: "../data"
 results: "../res"
 lib: "../lib"
 maxcores: 6
output:
 html_notebook:
 theme: spacelab
 highlight: textmate
 toc: yes
 number_sections: true
 toc_depth: 3
 toc_float: true
---

# using the package pcaExplorer and clusterprofiler
you need to use the data with the symbol and entrez
https://bioconductor.statistik.tu-dortmund.de/packages/3.6/bioc/vignettes/clusterProfiler/inst/doc/clusterProfiler.html#dotplot 

https://bioconductor.org/packages/devel/bioc/vignettes/pcaExplorer/inst/doc/pcaExplorer.html#9_Functions_exported_by_the_package_for_standalone_usage 
```{r}
library(DESeq2)
library(clusterProfiler)
BiocManager::install("airway")
library(airway)
data(airway)

dds_airway <- DESeqDataSet(airway,design= ~ cell + dex)
#dds_airway<-DESeqDataSetFromMatrix(countData = PAS,
 #                           colData=anno,
  #                          design = ~ sectio+ Csec+ Grav + Para+  GWscaled +place)
#adding an info about the ENS code

rld_airway <- rlogTransformation(dds_airway)
rld_airway
```

##On airway data 

```{r}
#pcaExplorer(dds = dds_airway,
            #dst = rld_airway)
library(org.Hs.eg.db)
keytypes(org.Hs.eg.db)

# didnt work with org.HS. trying with ENSsDb
if (!require("BiocManager", quietly = TRUE))
    install.packages("BiocManager")

#BiocManager::install("EnsDb.Hsapiens.v75")
library(EnsDb.Hsapiens.v75)
library(pcaExplorer)

dds_airway <- DESeq(dds_airway)
res_airway <- results(dds_airway) #this will be used from dds object from 1.0 metadata
genenames_airway <- mapIds(org.Hs.eg.db,keys = rownames(dds_airway),column = "SYMBOL",keytype="ENSEMBL")
annotation_airway <- data.frame(gene_name = genenames_airway,
                                row.names = rownames(dds_airway),
                                stringsAsFactors = FALSE)
head(annotation_airway)   

res_airway$symbol <- mapIds(org.Hs.eg.db,
                            keys= row.names(res_airway),
                            column="SYMBOL",
                            keytype="ENSEMBL",
                            multiVals="first")
res_airway$entrez <- mapIds(org.Hs.eg.db, ##problem 
                            keys=row.names(res_airway),
                            column="ENTREZID",
                            keytype="ENSEMBL",
                            multiVals="first")
resOrdered <- as.data.frame(res_airway[order(res_airway$padj),])
head(resOrdered)
# extract DE genes
de_df <- resOrdered[resOrdered$padj < .05 & !is.na(resOrdered$padj),]
de_symbols <- de_df$symbol
# extract background genes
bg_ids <- rownames(dds_airway)[rowSums(counts(dds_airway)) > 0]
bg_symbols <- mapIds(org.Hs.eg.db,
                     keys=bg_ids,
                     column="SYMBOL",
                     keytype="ENSEMBL",
                     multiVals="first")
# run the function
topgoDE_airway <- topGOtable(de_symbols, bg_symbols,
                             ontology = "BP",
                             mapping = "org.Hs.eg.db",
                             geneID = "symbol")

```
# on PAS data

https://bioconductor.org/packages/devel/bioc/vignettes/pcaExplorer/inst/doc/pcaExplorer.html#97_topGOtable 

```{r}
#using res, anno, PAS 
dds_GO <- DESeq(dds)
res<- results(dds_GO)
ENS_res<- merge(as.data.frame(res), KEY, by.y="gene_id", by.x= "row.names")
ENS_res$Row.names<-gsub("\\..*","",ENS_res$Row.names)
rownames(ENS_res)<-make.names(ENS_res$Row.names, unique=T) #renaming the rownames
#names(ENS_res)[names(ENS_res) == "Row.names"] <- "gene_id"

genenames_GO<-mapIds(org.Hs.eg.db,
                     keys=ENS_res$Row.names, 
                     column = "SYMBOL",
                     keytype = "ENSEMBL")
annotation_GO <- data.frame(gene_name = genenames_GO,  
                                row.names = row.names(ENS_res),
                                stringsAsFactors = FALSE)
head(annotation_GO)  

#anno_df_orgdb <- get_annotation_orgdb(dds = dds_GO,
                                      #orgdb_species = "org.Hs.eg.db",
                                      #idtype = "ENSEMBL")

#anno_df_biomart <- get_annotation(dds = dds_GO,
                                  #biomart_dataset = "hsapiens_gene_ensembl",
                                  #idtype = "ensembl_gene_id")

intersect(rownames(res),keys(org.Hs.eg.db,"ENSEMBL")) # not all keys are in the rownames


res$symbol<- mapIds(org.Hs.eg.db,
                            keys= row.names(ENS_res),
                            column="SYMBOL",
                            keytype="ENSEMBL",
                            multiVals="first")
res$entrez <- mapIds(org.Hs.eg.db, ##problem 
                            keys=row.names(ENS_res), #these keys are not correct
                            column="ENTREZID",
                            keytype="ENSEMBL",
                            multiVals="first")
resOrdered <- as.data.frame(res[order(res$padj),])
head(resOrdered)

de_df <- resOrdered[resOrdered$padj < .05 & !is.na(resOrdered$padj),]
de_symbols <- de_df$symbol
# extract background genes
bg_ids <- rownames(dds_GO)[rowSums(counts(dds_GO)) > 0]
bg_ids<-gsub("\\..*","",bg_ids)
bg_symbols <- mapIds(org.Hs.eg.db,
                     keys=bg_ids,
                     column="SYMBOL",
                     keytype="ENSEMBL",
                     multiVals="first")
# run the function
topgoDE_GO <- topGOtable(de_symbols, bg_symbols,
                             ontology = "BP",
                             mapping = "org.Hs.eg.db",
                             geneID = "symbol")
#working on 11.7.2022

```

##ReactomePA

[ ReactimePA](https://bioconductor.org/packages/release/bioc/manuals/ReactomePA/man/ReactomePA.pdf)
[vignette](https://yulab-smu.top/biomedical-knowledge-mining-book/reactomepa.html)

Input data is a vector of Gene IDs

```{r reactomePA}
library(ReactomePA)
library(DOSE)

#converting the ensebl to entrez id
library("AnnotationDbi")
library("org.Hs.eg.db")
ENS_res$ENTREZID_id = mapIds(org.Hs.eg.db,
                       keys=ENS_res$Row.names, 
                        column="ENTREZID",
                        keytype="ENSEMBL",
                        multiVals="first")

#creating an object of ENTREZID_ids and padjust
Gene_ID<-ENS_res %>%
  as.tibble() %>%
  dplyr::select(c(ENTREZID_id, padj))

gene_nm<-Gene_ID$ENTREZID_id
GeneID <- Gene_ID$padj #this is not correct 
#GeneID<-as.numeric(GeneID) # converting doesnt work as.numeric
names (GeneID) <- gene_nm #naming the vector 
head(GeneID)

#For the enrichPathway you need EntrezID- therefore you take gene_nm
R_PA<-enrichPathway(gene=gene_nm, pvalueCutoff = 0.05, readable = T)
head(R_PA)

## Reactome pathway gene set enrichment analysis
GeneID<- sort(GeneID, decreasing = T)
R_pathway<- gsePathway(GeneID,
                       pvalueCutoff = 0.2,
                       pAdjustMethod = "BH",
                       verbose=F)

# trying according to vignette--> doesnt work 
data(geneList, package="DOSE"
)
de <- names(geneList)[abs(geneList) > 1.5]
head(de)
de<-sort(de, decreasing = T)
R_pathway<- gsePathway(de,
                       pvalueCutoff = 0.2,
                       pAdjustMethod = "BH",
                       verbose=F)
```

## Visualisation of the pathways 
```{r Visualisation of the pathways }
BiocManager::install("graphite")
library("graphite")

viewPathway("PAS invasion vs. control", 
            readable = TRUE, 
            foldChange = GeneID)
## Error message: "unable to find an inherited method for function 'convertIdentifiers' for signature '"NULL"
##Comment: got this even when working with vignette data 
```
