==> analysis/5.0_interactions.rmd <==
---
title: "5.0_interactions"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
 
# Interactions between place and Csec

```{r}
anno<- readRDS("anno.rds")
PAS<-readRDS("PAS.rds")
dds<-DESeqDataSetFromMatrix(countData = PAS,
                            colData=anno,
                            design = ~ Grav + Para + sectio + place + sectio:place)
keep<-rowSums(counts(dds)) >=10
dds<-dds[keep,]

```

## PAS vs ctrl 
```{r}

#<!-- VJM: I would run this only once and save the output to a file. This saves you running intensive tasks multiple times, makes your script more readable and avoids changing the input varaibles in one place and forgetting to change it on the other places.  -->

dds$place<-factor(dds$place, levels= c("ctrl", "mus", "PAS"))
dds <- DESeq(dds)
results_PAS_ctrl <- results(dds)
results_PAS_ctrl<-results(dds, name="place_PAS_vs_ctrl")
results_PAS_ctrl<-results(dds, filterFun= ihw, contrast = c("place", "PAS", "ctrl"))

results_PAS_ctrl<- as.data.frame(results_PAS_ctrl)
results_PAS_ctrl<-results_PAS_ctrl[order(results_PAS_ctrl$padj),]
sig.res<-results_PAS_ctrl|>  #just the sign
  subset(padj<0.05)
# export the rds or csv
```
## mus vs. ctrl
```{r}

results_mus_ctrl<-results(dds, filterFun= ihw, contrast = c("place", "mus", "ctrl"))
results_mus_ctrl<- as.data.frame(results_mus_ctrl)
results_mus_ctrl<- results_mus_ctrl %>%
  subset(padj<0.05)
```

## pas vs mus
```{r}
results_mus_PAS<-results(dds, filterFun= ihw, contrast = c("place", "mus", "PAS"))
results_mus_PAS<- as.data.frame(results_mus_PAS)
results_mus_PAS<- results_mus_PAS %>%
  subset(padj<0.05)
```

```{r}
ENS_sig_int<- merge(as.data.frame(sig.res), PASbackup, by.y="gene_id", by.x= "row.names")
ENS_bg_int<-merge(as.data.frame(results_mus_ctrl), PASbackup, by.y="gene_id", by.x= "row.names")
```


```{r}


dds$group<- factor(paste0(dds$place, dds$sectio))
#design(dds)<- -group
dds<- DESeq(dds)
resultsNames(dds)
results(dds, name="place_PAS_vs_ctrl")
results(dds, contrast= c("place", "PAS", "ctrl" )) #how to check for other interactions
?results

```
