==> analysis/6.0_var_part.rmd <==
---
title: "Var_part.rmd"
author: Tina Gorsek Sparovec
date: february 2022
params:
data: "../4.0_data"
results: "../2.0_results"
lib: "../0.0_lib"
maxcores: 6
output: html_notebook
theme: spacelab
highlight: textmate
toc: yes
number_sections: true
toc_depth: 3
toc_float: true
---


# Variance partitioning 

[variance Partition workflow](https://r.qcbs.ca/workshop10/book-en/variation-partitioning.html)
[variancePartition BioCond vignette](https://www.bioconductor.org/packages/devel/bioc/vignettes/variancePartition/inst/doc/variancePartition.pdf)

framework for understanding drivers of variation
in gene expression in experiments with complex designs. A
typical application would consider a dataset of gene expression from individuals
sampled in multiple tissues or multiple time points where the goal is to understand variation within versus between individuals and tissues. variancePartition
use a linear mixed model to partition the variance attributable to multiple variables in the data. 


Needs normalised expression values

3 components:
1) gene expression matrix
2) metadata
3) model
```{r, include=FALSE}
#if (!require("BiocManager", quietly = TRUE))
 #   install.packages("BiocManager")

#BiocManager::install("variancePartition")
library(variancePartition)
library(limma)
library(BiocParallel)
library(DESeq2)
library(tidyverse)

#data are sourced from analysis/1.0 metadata

sig.res 
anno 
```
## From the vignette

```{r example}
#example
data(varPartData)
form<-~Age+ (1|Individual) + (1| Tissue) + (1| Batch)a
varPart<- fitExtractVarPartModel( geneExpr, form, info)
vp<-sortCols(varPart)
plotPercentBars(vp[1:10,])
plotVarPart(vp)
```

## Variance partition PAS

Working on the rlog results
```{r}
dds<-estimateSizeFactors(dds) 
expr <- rowSums(assay(rlog(dds))>1) >= 0.5* ncol(dds)

#transformation, plus filter for the genes that pass the 0.05
quantlog<-log2(assay(rlog(dds))[expr,] +1)

#changing the ENSEMBL to Gene names

quantlog <-  merge(as.data.frame(quantlog), 
                 KEY, 
                 by.y="gene_id", 
                 by.x= "row.names")
quantlog <-quantlog[!duplicated(quantlog$gene_name),]
quantlog <-  quantlog |>
  select(-Row.names)|>
  remove_rownames()|>
  column_to_rownames(var ="gene_name")

form <- ~( Grav + Para + Csec+ place + GWscaled) #which one is categorical, fixed or continuous
#uses rld table results for input
varPart_pas <- fitExtractVarPartModel(quantlog,form, anno) #problem
plotVarPart(varPart_pas)

```

```{r}
# plot of varienace fractions for a subset of genes
plotPercentBars(varPart_pas[1:10,], col = c(ggColorHue(ncol(varPart_pas) - 1), "grey85"))

```
How to get the genes not the ensemble ID? 

<!-- VJM: Use the KEY you generated and merge it ;)  -->

```{r}
sortCols(varPart_pas)
plotVarPart(sortCols(varPart_pas ) )
```
