==> analysis/7.0_visualisation.rmd <==
---
title: "7.0_visualisation"
author: Tina Gorsek Sparovec
date: 24.2.22
params:
 data: "../data"
 results: "../res"
 lib: "../lib"
 maxcores: 6
output:
 html_notebook:
 theme: spacelab
 highlight: textmate
 toc: yes
 number_sections: true
 toc_depth: 3
 toc_float: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(pasilla)
library(DESeq2)
library(tidyverse)
library(IHW)
library(pheatmap)
library(genefilter)
library(RColorBrewer)
library(vsn)
library(ggrepel)
library(ggnewscale)
#devtools::install_github("kevinsblake/NatParksPalettes", force = TRUE)
library(NatParksPalettes)
library(viridisLite)
library(viridis)
library(VennDiagram)

conflict_prefer("rowVars", "genefilter")

# Required data:

# (1) rld normalisation

rld

# (2) KEY - df of df genes with gene_name and gene_id

KEY

# (3) dds object, created in the 1.0_metadata_exploration

dds


```
## Venns diagram of genes
```{r}
#results_mus_ctrl 
#results_mus_PAS
#results_PAS_ctrl
PAS_ven <- rownames(sig.res)
MUS_ven <- rownames(results_mus_PAS)
ctrl_ven <-rownames(results_mus_ctrl)
pdf(file = file.path(params$res,"venn_diagram.pdf"))
venn.plot <- venn.diagram(list(PAS_ven, ctrl_ven, MUS_ven), filename = NULL,
                          category = c("PAS", "ctrl", "MUS"))
                                
grid.draw(venn.plot)
dev.off()
```


```{r normalisation}
rld <- rlog(dds, blind=FALSE)

```

## Sample distances

aim: assesing the overall similarity between the samples- which samples are similar to each other

```{r}
sampleDists <- dist(t(assay(rld)))

sampleDistMatrix <- as.matrix(sampleDists)

rownames(sampleDistMatrix) <- paste(rld$place, 
                                    anno$patient,
                                    sep="-")
colnames(sampleDistMatrix) <- NULL
colors <- c(magma(20))
pheatmap(sampleDistMatrix,
         clustering_distance_rows=sampleDists,
         clustering_distance_cols=sampleDists,
         col=colors)
```
[] use the IDs for the y axis

## Gene clustering/heatmap with dendogram

[heatmap](http://bioconductor.org/help/course-materials/2015/BioC2015/bioc2015rnaseq.html)
[from pheatmap to complexheatmap](https://jokergoo.github.io/2020/05/06/translate-from-pheatmap-to-complexheatmap/)
```{r}

#selecting 20.000 genes for the heatmap
topVarGenes <- head(order(-rowVars(assay(rld))), 20000) 

mat <- assay(rld)[ topVarGenes, ]
mat <- mat - rowMeans(mat)
mat <- mat |>
  as.data.frame() |>
  rownames_to_column(var= "Ensembl")

# merging with KEY to get the gene_name 
mat <- merge(mat,
             KEY, 
             by.y="gene_id", 
             by.x= "Ensembl")

# removing the duplicated genes and ordering of the columns
order <- c("TG.1", "TG.10", "TG.2", "TG.3", "TG.4", "TG.6", "TG.7", "TG.9", "TG.12", "TG.14", "TG.15", "TG.17", "TG.18", "TG.20", "TG.21", "TG.13", "TG.16", "TG.19")
mat <- mat[!duplicated(mat$gene_name),]
mat <- mat |>
  select(-Ensembl) |>
  remove_rownames()|>
  column_to_rownames(var= "gene_name")  |> 
  relocate("TG.1", "TG.10", "TG.2", "TG.3", "TG.4", "TG.6", "TG.7", "TG.9", "TG.12", "TG.14", "TG.15", "TG.17", "TG.18", "TG.20", "TG.21", "TG.13", "TG.16", "TG.19")

# selecting the cluster genes
#cluster 1 and cluster 3
matclust_3 <- mat |> 
  filter(row.names(mat) %in% c(cluster_3_genes$gene_name[1:25]))
matclust_1 <- mat |> 
  filter(row.names(mat) %in% c(cluster_1_genes$gene_name[1:25]))

# annotate the heatmap
df <- as.data.frame(colData(rld)[,c("place", "patient")])
df <- df |> 
  select(place) |> 
  arrange(place)


#col <- colorRampPalette(rev(brewer.pal(n = 7, name ="RdYlBu")))
#pat_col <-natparks.pals("Arches2", 15)
#names(pat_col)<-levels(df$patient)

ann_col <- list(place = c(Ctrl = "#fd9668", PAS = "#9e2f7f", CtrlMus = "#440f76"))

pdf(file = file.path(params$res, "heatmap_clust.pdf"),width = 10, height = 10)
 ph3 <- pheatmap(matclust_3, 
         annotation_col = df,
         color = magma(10),
         annotation_colors = ann_col,
         annotation_names_col = T,
         border_color = NA,
         cluster_cols = FALSE, 
         main = "Upregulated in PAS")
 ph1 <- pheatmap(matclust_1, 
         annotation_col = df,
         color = magma(10),
         annotation_colors = ann_col,
         annotation_names_col = T,
         border_color = NA,
         legend = FALSE,
         main = "Downregulated genes in PAS",
         cluster_cols = FALSE,
         annotation_legend = FALSE)
 plot_list <- list()
plot_list[["ph1"]] <- ph1[[4]]
plot_list[["ph3"]] <- ph3[[4]]
gridExtra::grid.arrange(grobs = plot_list, ncol =2)
   
       dev.off() 

```
[] try to plot the genes of the cluster 3
[] you can make this for the GO results
[] instead of place/pat you can use the GO term- add to metadata



## vulcano plots


```{r}

# assigning the genes to the column 
summary(results_PAS_ctrl, alpha = 0.05)

# creating a tibble from results 
results_PAS_ctrl2 <- results_PAS_ctrl|>
   data.frame() |>
   rownames_to_column(var="gene_id") |>
   as_tibble() |>
   mutate(threshold_OE = padj < 0.05 & abs(log2FoldChange) >= 0.58) |>
   mutate(gene_label="")

# merging with KEY to get the gene_names
results_PAS_ctrl2 <-merge(as.data.frame(results_PAS_ctrl2), 
                 KEY, 
                 by.y="gene_id", 
                 by.x= "gene_id") 

# arranging padj
results_PAS_ctrl2 <- results_PAS_ctrl2 |>
  arrange(padj) 

#naming first genes in a separate column
#results_PAS_ctrl2$gene_label <- results_PAS_ctrl2$gene_name


## Kegg anaylsis genes: TGF-beta signaling pathway
tgf <- c(	"ID4", "INHBA", "SMAD3", "PITX2", "ID1", "GDF7", "FBN1", "ACVR1C", "TGFB2", "TGIF1")

## reactome analysis
ECM_proteoglycans <- c("COL4A5", "SERPINE1", "COL4A1", "BGN", "TGFB2", "TNXB", "VCAN")

## R_PA analysis genes:
#Signaling by TGFB family members
sig_TGFB_path <- c( "CHRDL1","SERPINE1", "PMEPA1", "INHBA", "SMAD3", "FBN1", "ACVR1C", "FSTL1", "TGFB2", "FURIN", "TGIF1")


# Cluster 1 genes 
results_PAS_ctrl2$genelabels <- factor(results_PAS_ctrl2$gene_name, levels = sig_TGFB_path) #row.names(matclust_1))

pdf(file = file.path(params$res, "vulcano_tgfb.pdf"),width = 6, height = 5)
ggplot(results_PAS_ctrl2, 
       aes(x = log2FoldChange, y = -log10(padj))) +
    geom_point(aes(colour = threshold_OE), size = 0.3, show.legend = FALSE) +
    geom_text_repel(aes(label= genelabels), max.overlaps = Inf) +
    ggtitle("PAS overexpression") +
    xlab("log2 fold change") + 
    ylab("-log10 adjusted p-value") +
    theme(legend.position = "none",
         plot.title = element_text(size = rel(1.5), hjust = 0.5),
         axis.title = element_text(size = rel(1.25))) +
    #scale_fill_discrete(guide="none")+
    scale_color_manual(values= c( "#ca3e72", "#feca8d")) +
    theme_classic()
dev.off()
#[]adjust to maybe just top 10 DEG in clust 1 and 3
       
```
[] label just the most interesting genes

? is it a good idea to label only the genes that would be of interest?
## Biological theme comparison based on clusters 

```{r}
#### move this to the top


# For this analysis you need a list of genes created with degPatterns, 1.0_metadata_exploratioj

clust_df <- clusters$df
clusters$plot

#adding a column with ENTREZID
clust_df$ENTREZID <- mapIds(org.Hs.eg.db,
                       keys=clust_df$genes, 
                        column="ENTREZID",
                        keytype="ENSEMBL",
                        multiVals="first")


#extracting a list with genes of each cluster
clust <- c(1:4)
list<-list()
for(i in 1:length(clust)) {
c<- clust_df |>
  rownames_to_column(var = "gene_id") |>
  left_join(KEY) |>
  relocate(gene_name, .after = gene_id) |>
  filter(cluster == i) |>
  select(-genes) |> 
  pull(gene_name)
  
list[[i]]<- c
}

list <- lapply(list, function(x) x[!is.na(x)]) #omiting NA values




```
## heatmap with GO terms
```{r}
# https://d-nb.info/1202186009/34 

#selecting annotation_col
dfx <- as.data.frame(colData(rld))
dfx <- dfx |> 
  select(place)

### selected genes
##Kegg pathway analysis
tgf <- c(	"ID4", "INHBA", "SMAD3", "PITX2", "ID1", "GDF7", "FBN1", "ACVR1C", "TGFB2", "TGIF1")
mat_tgf <- mat |> 
  filter(row.names(mat) %in% tgf)

##BP
cell_migration <- c("ARX","FYN","HGF","WWC3","HEXB","PIK3CB","KIF2A","GPC4","B4GALT1","PXN","LAMB4","TGFB2","XBP1","SFRP1","PLAT","PTN","SERPINE1","PDLIM1","PDGFRB","IGFBP5","TNFAIP3","PTK2B","CXCR4","ACVR1C","SDC4","TNFSF14","ID1","PODXL","SWAP70","NAV1","ANXA1","CGA","TUBB2A","PRCP","CYP19A1","ANXA3","IQSEC1","PHLDB2","SH3KBP1","GPC3","PSTPIP2","FAM49B","ROBO4","APCDD1","FCER1G","STC1","ACE","S100A11","FSTL1","CCDC141","PITX2","RET","NDEL1","SMAD3","IGFBP6","ULK4","SEMA4C","ATOH8","HAS2","MCC","FZD4","EPHB3","PIP5K1C","DNER","APOD","PPIA","CDC42BPB",
"LRIG2","DDR1","PLXNA4","PDCD6","S1PR2")
mat_cell_mig <- mat |> 
  filter(row.names(mat) %in% cell_migration)

neg_reg_growth <- c("JARID2","WWC3","TGFB2","SFRP1","IGFBP5","INHBA","SH3BP4","CGA","GPC3","SMAD3","GNG4","SEMA4C","MYOZ1","SERTAD2","WT1","MT1X")
mat_neg_reg_growth <- mat |> 
  filter(row.names(mat) %in% neg_reg_growth)

cell_matrix_adh <- c("PIK3CB","SFRP1","SERPINE1","PTK2B",
"SDC4","PHLDB2","UTRN","SMAD3",
"APOD","DDR1")
mat_cell_matrix <- mat |> 
  filter(row.names(mat) %in% cell_matrix_adh)

cell_cell_adh <- c("FYN","PIK3CB","GPC4","TGFB2","XBP1","PDLIM1","ICAM2","PERP","PCDH17","SDC4","TNFSF14","CDC42EP1","PODXL","FLOT2","SWAP70","DSC2","DSC3","ANXA1","ADAM19","PCDH10","SLC7A1","RPS3","FAM49B","ROBO4","S100A11","NFASC","RET","TNXB","HAS2","LGALS9C","CEBPB","EPHB3","PIP5K1C","PPIA","TENM3")
mat_cell_cell <- mat |> 
  filter(row.names(mat) %in% cell_cell_adh)

##MF
#MF involved in binding
bind <- c("GPRC5A","SLC9A3R2","NOTCH3","COBLL1","TBC1D10A","NDRG1","PPP1R13L","PTN","PDLIM1","ICAM2","EIF4G2","ADAM23","SPTBN1","VASP","CDC42EP1","SWAP70","DSC2","ANXA1","BZW2","USO1","PHLDB2","UTRN","ROBO4","TAGLN2","S100A11","NFASC","FBN1","TNXB","BCAM","PPIA","EGFL6","TENM3","S1PR2", # cell adhesion moecule binding; cadherine binding GO:0045296, cadherin binding involved in cell-cell adhesion GO:0098641 #integrin binding GO:0005178
          "IGFBP5","SDC4","IGFBP6", #fibronectin binding
          "VCAN","SFRP1","STAB2","PCSK6","HDGF","FSTL1","NAV2","BGN","SULF2", #glycosaminoglycan binding GO:0005539
          "APCDD1","FZD4", #wnt-protein binding GO:0017147
          "DDX5","PMEPA1","SMAD3" # R-smad binding GO:0070412
         )
### help here please
mat_bind <- mat |> 
  filter(row.names(mat) %in% bind) |> 
  arrange(bind)
###   

# MF for extra cellular matrix
ECM <- c("VCAN","OGN","BGN")
mat_ECM <- mat |> 
  filter(row.names(mat) %in% ECM)


#makeHeatmap function
makeHeatmap <- function (selected_genes, name_of_heatm, annotation_col) {
  hm <- pheatmap (selected_genes, 
                  annotation_col= annotation_col,
                  color = magma(90),
                  annotation_colors = ann_col,
                  annotation_names_col = T,
                  border_color = NA,
                  cluster_cols = FALSE,
                  cluster_rows = FALSE, #is this something that we need?
                  main = name_of_heatm)
  return(hm)
}

heatmap_tgf <- makeHeatmap(mat_tgf, "TGF signalling pathway", dfx)
pdf(file = file.path(params$res, "heatmap_cell_mig.pdf"),width = 10, height = 10)
heatmap_cell_mig <- makeHeatmap(mat_cell_mig, "GO cell migration", dfx)
dev.off()
heatmap_neg_reg_growth <- makeHeatmap(mat_neg_reg_growth, "GO negative regulation of growth", dfx)
heatmap_cell_matrix <- makeHeatmap(mat_cell_matrix, "GO cell-matrix adhesion", dfx)
heatmap_cell_cell <- makeHeatmap(mat_cell_cell, "GO cell-cell adhesion", dfx)
#pdf(file = file.path(params$res, "heatmap_binding.pdf"),width = 10, height = 10)
heatmap_bind <-makeHeatmap(mat_bind, "GO binding", dfx)
#dev.off()
```
To-do

[] create objects with genes
[] run the heatmap function for the selected objects
