==> analysis/8.0_immune.Rmd <==
---
title: "8.0_immune"
author: "Tina Gorsek"
date: '2022-07-25'
params:
 data: "../data"
 results: "../res"
 lib: "../lib"
 maxcores: 6
output:
 html_notebook:
 theme: spacelab
 highlight: textmate
 toc: yes
 number_sections: true
 toc_depth: 3
 toc_float: true
---
```{r, include=FALSE}
#install.packages("remotes")
#remotes::install_github("grst/immunedeconv")
```

```{r, message=FALSE, warning=FALSE}
library(tidyverse)
library(immunedeconv)
library(EPIC)
require(org.Hs.eg.db)

# objects needed for this analysis are

#(1) 
PAS

```

## Vignette

```{r}
dataset_racle <- immunedeconv::dataset_racle
dataset_racle$expr_mat

deconvolution_methods
```

## PAS data analysis

In order to analyse our data we need a gene expression matrix with HGNC symbols in rows 
```{r data wrangling}
PAS_immune <- PAS |>
        as.data.frame() |>
        rownames_to_column(var= "ENSEMBL") 

ensembl_to_entrez <- AnnotationDbi::select(org.Hs.eg.db, unique(PAS_immune$ENSEMBL),
                                           keytype = "ENSEMBL",
                                           columns = c("SYMBOL"))
PAS_immune <- PAS_immune |>
        inner_join(ensembl_to_entrez, by = c("ENSEMBL" = "ENSEMBL")) 
PAS_immune <- PAS_immune[!duplicated(PAS_immune$SYMBOL),]
PAS_immune <- PAS_immune |>
  as.data.frame() |>
        remove_rownames() |>
        drop_na() |>
        column_to_rownames(var= "SYMBOL")
PAS_immune <- PAS_immune |>
  dplyr::select(-ENSEMBL) |> 
  as.matrix()
order <- c("TG.1", "TG.10", "TG.2", "TG.3", "TG.4", "TG.6", "TG.7", "TG.9", "TG.12", "TG.14", "TG.15", "TG.17", "TG.18", "TG.20", "TG.21", "TG.13", "TG.16", "TG.19")
#write.table(PAS_immune, "PAS.txt", sep= " ")
```

## Deconvolution

```{r}
##methods:
#mpc <- "mcp_counter"
epic <- "epic"
#qseq <- "quantiseq"


res_cibersort <- deconvolute(PAS_immune, epic, tumor = FALSE)

```

```{r, fig.height=4, fig.width=8}
res_cibersort %>%
  gather(sample, fraction, -cell_type) %>%
  # plot as stacked bar chart
  ggplot(aes(x = sample, y = fraction, fill = cell_type)) +
  geom_bar(stat = "identity") +
  coord_flip() +
  scale_fill_brewer(palette = "Paired") +
  #scale_x_discrete(limits = rev(levels(res_cibersort))) +
  scale_x_discrete(limits = order) #reordering the y axis
```


```{r}
res_mcp_counter <- deconvolute(PAS_immune, epic, tumor = FALSE)
res_mcp_counter %>%
  gather(sample, score, -cell_type) %>%
  ggplot(aes(x = sample, y = score, color = cell_type)) +
  geom_point(size = 4) +
  facet_wrap(~cell_type, scales = "free_x", ncol = 3) +
  scale_color_brewer(palette = "Paired", guide = FALSE) +
  coord_flip() +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  scale_x_discrete(limits = order)
```
Idea: would be good to split the samples on PAS and ctrl for better visualisation 


selecting a method for the deconvolution:
- cibersort- measure intraeample proportions of immune cell population, MCP counter measure an abundance estimate per cell population that enables inter sample comparison
